Added ability to save data in app memory and added indxes of variables used
Added function that handle saving current day mood in app memory,as well as tracking if you change your mood in a day and delete same occurence of day in log. This system is provided with separate function that writes down all reesonable data of day when log is created. Also added variable indexes for getting day and emotion emoji.